fix: Resolve remaining analyzer errors (type conversions and icon)
- spending_evaluation_service.dart: Convert BudgetCategory to String (category.name) and use displayName in recommendations
- dashboard_provider.dart: Add .toInt() conversions for goal amount calculations
- category_breakdown_card.dart: Add .toDouble() for percentage calculation
- goals_progress_card.dart: Add type cast to double for progress calculation
- quick_actions_card.dart: Replace undefined Icons.target with Icons.track_changes

These changes resolve all 5 critical analyzer errors while maintaining functionality.

+++ lib/features/dashboard/presentation/providers/dashboard_provider.dart
@@ -72,8 +72,8 @@ final savingsDashboardProvider =
       int totalSaved = 0;
       int totalTarget = 0;
       for (final goal in goals) {
-        totalSaved += goal.currentAmount;
-        totalTarget += goal.targetAmount;
+        totalSaved += goal.currentAmount.toInt();
+        totalTarget += goal.targetAmount.toInt();
       }
       goalProgress = totalTarget > 0 ? totalSaved / totalTarget : 0.0;
     }
